+++ app/Livewire/Navigation.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire;
+
+use Livewire\Component;
+
+class Navigation extends Component
+{
+    public function render()
+    {
+        return view('livewire.layout.navigation');
+    }
+}
