+++ bootstrap/providers.php
@@ -3,5 +3,4 @@
 return [
     App\Providers\AppServiceProvider::class,
     App\Providers\HorizonServiceProvider::class,
-    App\Providers\VoltServiceProvider::class,
 ];
