+++ composer.json
@@ -19,7 +19,6 @@
         "laravel/sanctum": "^v4.0",
         "laravel/tinker": "^v2.10",
         "livewire/livewire": "^v3.6",
-        "livewire/volt": "^v1.6",
         "maatwebsite/excel": "^3.1",
         "openai-php/laravel": "^v0.11",
         "predis/predis": "^v2.3",
