+++ resources/views/livewire/navigation.blade.php
@@ -1,0 +1,3 @@
+<div>
+    {{-- Success is as dangerous as failure. --}}
+</div>
